+++ learn-go-with-tests/02-build-an-application/InMemoryPlayerStore.go
@@ -1,0 +1,17 @@
+package main
+
+func NewInMemoryPlayerStore() *InMemoryPlayerStore {
+    return &InMemoryPlayerStore{map[string]int{}}
+}
+
+type InMemoryPlayerStore struct{
+	store map[string]int
+}
+
+func (i *InMemoryPlayerStore) RecordWin(name string) {
+    i.store[name]++
+}
+
+func (i *InMemoryPlayerStore) GetPlayerScore(name string) int {
+    return i.store[name]
+}
