+++ learn-go-with-tests/02-build-an-application/main.go
@@ -1,0 +1,12 @@
+package main
+
+import (
+	"log"
+	"net/http"
+)
+
+func main() {
+    server := &PlayerServer{NewInMemoryPlayerStore()}
+    log.Fatal(http.ListenAndServe(":5000", server))
+}
+
