+++ learn-go-with-tests/02-build-an-application/server_test.go
@@ -104,3 +104,14 @@ func newPostWinRequest(name string) *http.Request {
     req, _ := http.NewRequest(http.MethodPost, fmt.Sprintf("/players/%s", name), nil)
     return req
 }
+
+func TestLeague(t *testing.T) {
+    store := StubPlayerStore{}
+    server := &PlayerServer{&store}
+    t.Run("it returns 200 on /league", func(t *testing.T) {
+        request, _ := http.NewRequest(http.MethodGet, "/league", nil)
+        response := httptest.NewRecorder()
+        server.ServeHTTP(response, request)
+        assertStatus(t, response.Code, http.StatusOK)
+    })
+}
